index: use SSH key authentication if ssh_privkey_path is set
Previously, if the USER_TYPE_PLAINTEXT credential type was supported,
then a provided SSH key would never be used even if no HTTPS config
options were supplied. After this change, if `ssh_privkey_path` is set
and the SSH_KEY credential type is supported, then we attempt
to use the private key. The HTTPS username and password are only used if
`ssh_privkey_path` is not set or the SSH_KEY credential type is not
supported.